Criação das rotas de navios e realizados testes com 1 e 2 berços, com 1 e 2 portainers cada um deles, com diversos parâmetros de tempo para validação.

+++ Simulador/src/simulador/random/UniformDistributionStream.java
@@ -20,7 +20,7 @@ public class UniformDistributionStream implements DistributionFunctionStream{
         
     @Override
     public double getNext() {
-        return _aleatorio.getNext();
+        return (int)_aleatorio.getNext();
     }
     
 }
